DB_usage_in_python/requirement.txt
pillow==11.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
